+++ pkg/kiosk/utils.go
@@ -26,20 +26,26 @@ func GenerateURL(cfg *Config) string {
 	case "full": // FULLSCREEN
 		parsedQuery.Set("kiosk", "1") // sidebar and topnav always shown
 		log.Printf("KioskMode: Fullscreen")
-	case "disabled": // FULLSCREEN
+	case "disabled":
 		log.Printf("KioskMode: Disabled")
 	default: // disabled
 		parsedQuery.Set("kiosk", "1") // sidebar and topnav always shown
 		log.Printf("KioskMode: Fullscreen")
 	}
 	if cfg.General.HideLinks {
-		parsedQuery.Set("_dash.hideLinks", "")
+		parsedQuery.Set("_dash.hideLinks", "true")
+	}
+	if cfg.General.HideLogo {
+		parsedQuery.Set("_dash.hideLogo", "1") // Grafana uses "1" for this param, not "true"
+	}
+	if cfg.General.HidePlaylistNav {
+		parsedQuery.Set("_dash.hidePlaylistNav", "true")
 	}
 	if cfg.General.HideTimePicker {
-		parsedQuery.Set("_dash.hideTimePicker", "")
+		parsedQuery.Set("_dash.hideTimePicker", "true")
 	}
 	if cfg.General.HideVariables {
-		parsedQuery.Set("_dash.hideVariables", "")
+		parsedQuery.Set("_dash.hideVariables", "true")
 	}
 	// a playlist should also go inactive immediately
 	if cfg.Target.IsPlayList {